requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
openai
langchain
streamlit
langchain.chains

